add internalBucketInfo
